requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
apptools==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autograd==1.3
autograd-gamma==0.5.0
backcall==0.2.0
bleach==4.1.0
Braindecode==0.6
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
configobj==5.0.6
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dipy==1.4.1
entrypoints==0.4
envisage==6.0.1
fonttools==4.28.3
formulaic==0.2.4
frozenlist==1.2.0
future==0.18.2
h5py==3.6.0
hdf5storage==0.1.18
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
interface-meta==1.2.4
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lifelines==0.26.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mayavi==4.7.4
mistune==0.8.4
mne==0.24.1
multidict==5.2.0
natsort==8.0.2
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nibabel==3.2.1
nilearn==0.8.1
notebook==6.4.8
numpy==1.21.4
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.2
pooch==1.5.2
prettytable==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pycparser==2.21
pyface==7.3.0
Pygments==2.10.0
pyparsing==3.0.6
#PyQt5==5.15.6
#PyQt5-Qt5==5.15.2
#PyQt5-sip==12.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
#QtPy==2.0.1
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
skorch==0.11.0
stg==0.1.2
tabulate==0.8.9
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
traits==6.3.2
traitsui==7.2.1
typing_extensions==4.0.1
urllib3==1.26.7
vtk==9.0.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
widgetsnbextension==3.5.2
wrapt==1.13.3
wslink==1.2.0
yarl==1.7.2
zipp==3.6.0
